Add classes to wrap Phasor native code
This commit adds two different files.
GCI_Phasor.i:
This is the SWIG interface file. It declares
functions which are to be wrapped and also
global variables and constants.
SwigWrapperTest.java:
This class tests to make sure the wrapped
code can be accessed in java via the
NarSystem loading mechanism.