Add files to wrap EcfGlobal native code
This commit adds two different files.
EcfGlobal.i:
This is the SWIG interface file. It declares
functions which are to be wrapped and also
global variables and constants. Ecf.h was
included to deal with enums declared there.
EcfGlobal.h:
Simplifies the interface file. Simply declares
the functions to be wrapped in EcfGlobal.c