Enable ptr to ptr and function ptr params
This commit enables pointer to pointers to be wrapped
in swig code.  In java a simple array is passed in as an
argument. Function pointers can also be wrapped now; however,
java functions can not be passed in as arguments.  Instead, a
c function has to be declared as being a potential function
which gets pointed to and this can be wrapped.

By implementing these changes, the need for using SWIGTYPES
was removed.  This greatly improves the readability of the java
code which calls the native library.