Make license update ignore vendered source files
Unfortunately, the license-maven-plugin is finicky when it comes to
match patterns. In particular, there seems to be a bug when matching the
first directory beneath the project root. The following do *not* work:

* src/flimlib-cmd/c/initparser/**
* /src/flimlib-cmd/c/initparser/**
* **/src/flimlib-cmd/c/initparser/**

But these patterns *do* work:

* **/flimlib-cmd/c/initparser/**
* **/initparser/**

I opted for the most specific patterns, omitting only the /src prefixes.

Closes #77.