CI: let deploy job rebuild the Linux code
Now, we only drop in the macOS and Windows native artifacts and
attach those, rather than trying to do all of them including Linux,
and then forcing the deploy node to skip most of the Maven build.

The issue was that the deploy-to-scijava profile really wants to build
the javadocs, and building the javadocs requires the source code,
including the SWIG-generated source code... so we might as well just
let the deploy node go through the whole rigamarole. Yes, it means the
Linux code gets built twice. But the actual build is quite fast, so
whatever. We do only run the prysk tests on Linux once: on the build
node, not the deploy node.