CI: Don't release from a shallow clone copy
When the maven-release-plugin's perform goal runs, by default it does a
shallow clone from the release tag and builds from there, to ensure a
fresh space. But we don't actually want a fresh space here; we want the
native classifier JARs to be there. I tried dropping them into
`target/checkout/target`, but then the `git clone` command fails because
the `checkout` folder is not empty. Fortunately, the
maven-release-plugin supports skippiing this shallow clone step in favor
of just running the build+deploy from the already-correct working copy.